refs: make unborn-ness checking available outside of HEAD
The only place where GIT_EUNBORNBRANCH is returned is when a repository's
HEAD is looked up. It turns out there can be other symbolic branches
(eg. in worktrees), and they could be unborn as well, which we'd see as
GIT_ENOTFOUND.

Move the unbornness checking in a private ref-lookup function, so it can
be used for other cases.