clar: option to append summary file
Add a command line option `-R` that appends to the summary file instead
of overwriting it.  This allows users to run multiple clar invocations
and simply append to a single file, instead of trying to manage multiple
files.

Use a `fopen` / `fdopen` combination to open the file without
truncation, then ensure that either the file is empty or it ends with a
closing JUnit XML root node.  If it does, set the write cursor there and
omit the starting root node.

Since this complicates initialization, move the summary into the clar
object so that we can report errors back with the summary.