symlink tests: test symbolic links on windows
Test updated symbolic link creation on Windows.  Ensure that we emulate
Git for Windows behavior.  Ensure that when `core.symlinks=true` is set
in a global configuration that new repositories are created without a
`core.symlinks` setting, and that when `core.symlinks` is unset that
`core.symlinks=false` in set in the repository.  Further ensure that
checkout honors the expected `core.symlinks` defaults on Windows.